# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: P2P Game Item Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: Get current user's transactions
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Buy a listing (create transaction)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Bad request (insufficient balance, own listing, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/{transactionId}:
    get:
      operationId: getTransactionById
      tags: [transactions]
      summary: Get transaction by ID
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/{transactionId}/seller-confirm:
    post:
      operationId: sellerConfirmSent
      tags: [transactions]
      summary: Seller confirms item has been sent
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProofRequest"
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/{transactionId}/buyer-confirm:
    post:
      operationId: buyerConfirmReceived
      tags: [transactions]
      summary: Buyer confirms item has been received (releases escrow)
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProofRequest"
      responses:
        "200":
          description: Confirmed and escrow released
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/{transactionId}/dispute:
    post:
      operationId: disputeTransaction
      tags: [transactions]
      summary: Open a dispute for a transaction
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeRequest"
      responses:
        "200":
          description: Dispute opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        discordId:
          type: string
        avatarUrl:
          type: string
          nullable: true
        rating:
          type: number
          format: float
        totalTrades:
          type: integer
        joinedAt:
          type: string
          format: date-time
        walletBalance:
          type: number
          format: float
      required:
        - id
        - username
        - discordId
        - rating
        - totalTrades
        - joinedAt
        - walletBalance
...
    Listing:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        seller:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        game:
          type: string
        category:
          type: string
        price:
          type: number
          format: float
        imageUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, sold, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sellerId
        - seller
        - title
        - description
        - game
        - category
        - price
        - status
        - createdAt
        - updatedAt
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: integer
        listing:
          $ref: "#/components/schemas/Listing"
        buyerId:
          type: integer
        buyer:
          $ref: "#/components/schemas/User"
        sellerId:
          type: integer
        seller:
          $ref: "#/components/schemas/User"
        amount:
          type: number
          format: float
        status:
          type: string
          enum: [pending, awaiting_delivery, delivery_confirmed, completed, disputed, cancelled]
        sellerProofUrl:
          type: string
          nullable: true
        buyerProofUrl:
          type: string
          nullable: true
        disputeReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - listingId
        - listing
        - buyerId
        - buyer
        - sellerId
        - seller
        - amount
        - status
        - createdAt
        - updatedAt
    CreateTransactionRequest:
      type: object
      properties:
        listingId:
          type: integer
      required:
        - listingId
    ProofRequest:
      type: object
      properties:
        proofUrl:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - proofUrl
    DisputeRequest:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
...